requirements.txt
opencv-python
dlib
torch
numpy
librosa
hmmlearn
scikit-learn
transformers
pillow
cmake
dlib
numpy
opencv-python-headless
tensorflow
torch
transformers
pillow
face_recognition